GS-14610 Support arithmetic operations on multiple columns and literals Pass sqlTypeName as return type of function.
Note that for interval operations, the type represented additional info
for the function rather than the return value - thus the special handling.
Also, unit tests do not pass a sqlTypeName, if it is null we return default.

GS-14610 Fix arithmetic operations for interval

GS-14610 Fix NPE in unit tests that do not pass sqlTypeName

(cherry picked from commit 637469ca8e1b1ca522043c3c657435276b8b69fe)

+++ xap-core/xap-datagrid/src/main/java/com/gigaspaces/internal/utils/math/MutableFloat.java
@@ -33,7 +33,7 @@ public class MutableFloat extends MutableNumber {
 
     @Override
     public Number toNumber() {
-        return Float.valueOf(value);
+        return value;
     }
 
     @Override
